Refactor slice clip tooling into separate module
