+++ lib/fix-google-html.js
@@ -6,6 +6,11 @@ import { visitParents } from 'unist-util-visit-parents';
 import GithubSlugger from 'github-slugger';
 import debug from 'debug';
 import { resolveNodeStyle } from './css.js';
+import {
+  sliceClipText,
+  rangesForSuggestions,
+  replaceRangesInTree,
+} from './slice-clip.js';
 
 const log = debug('app:fix-google-html:debug');
 
@@ -72,7 +77,6 @@ const isSpaceSensitive = (node) =>
   node && spaceSensitiveElements.has(node.tagName);
 const isCell = (node) => node.tagName === 'th' || node.tagName === 'td';
 const isAnchor = (node) => node.tagName === 'a';
-const isText = (node) => node.type === 'text';
 const isSuggestion = (n) => n.properties?.['data-suggestion-id'] != null;
 
 const spaceAtStartPattern = /^(\s+)/;
@@ -664,127 +668,6 @@ function fixInternalLinks(node, sliceClip) {
 }
 
 /**
- * Visit each range of text as found in a HAST tree. You can replace or change
- * the node from within the visitor and it should continue to work.
- * @param {string} docsText
- * @param {GDocsRange[]} ranges
- * @param {RehypeNode} tree
- * @param {function} visitor
- */
-function forEachDocsRangeInTree(docsText, ranges, tree, visitor) {
-  if (ranges.length === 0) return;
-
-  ranges = ranges.slice().sort((a, b) => a.start - b.start);
-
-  let currentIndex = 0;
-  let currentNonspaceIndex =
-    currentIndex + docsText.slice(currentIndex).match(/^\s*/)[0].length;
-
-  visit(tree, isText, (node, nodeIndex, parent) => {
-    const leadingSpaces = node.value.match(/^\s*/)[0];
-    const text = node.value.slice(leadingSpaces.length);
-
-    const endIndex = currentNonspaceIndex + text.length;
-    const expectedText = docsText.slice(currentNonspaceIndex, endIndex);
-
-    if (expectedText !== text) {
-      console.warn(
-        `Iterating text did not match expected!\n  Range: ${currentIndex}...${endIndex}\n  Text: "${node.value}"\n  Expected: "${docsText.slice(currentIndex, endIndex)}"`
-      );
-      return EXIT;
-    }
-
-    if (ranges[0] && ranges[0].start < endIndex) {
-      let range = ranges.shift();
-      const startInText = range.start - currentNonspaceIndex;
-      const rangeStart = leadingSpaces.length + startInText;
-
-      // If the range extends into another node chop it up and put the
-      // remainder back on the list of ranges to handle.
-      const overreach = range.start + range.length - endIndex;
-      if (overreach > 0) {
-        const remainder = { ...range, start: endIndex, length: overreach };
-        range = { ...range, length: endIndex - range.start };
-
-        const point = ranges.findIndex((r) => r.start >= remainder.start);
-        if (point < 0) {
-          ranges.push(remainder);
-        } else {
-          ranges.splice(point, 0, remainder);
-        }
-      }
-
-      visitor({ node, nodeIndex, parent, range, rangeStart });
-
-      if (ranges.length === 0) {
-        return EXIT;
-      } else {
-        // The next range could also be in this node, so re-visit it.
-        return [CONTINUE, nodeIndex];
-      }
-    } else {
-      currentIndex = endIndex;
-      currentNonspaceIndex =
-        currentIndex + docsText.slice(currentIndex).match(/^\s*/)[0].length;
-    }
-  });
-}
-
-/**
- * @typedef {object} GDocsRange
- * @property {number} start
- * @property {number} length
- */
-
-/**
- * Create a set of range objects to represent the location of suggested changes
- * in the text based on data from a Slice Clip object.
- *
- * In the slice clip, ranges are represented by a large array, where each index
- * indicates a location in the text. If the value is an array, it will be an
- * array of suggestion IDs for suggestions that start (or continue) at that
- * location in the text. If a currently open suggestion is not listed in the
- * value, the location marks the end of that suggestion.
- *
- * For exampled, suggestion 'a' starts at index 2 and ends at 7; 'b' starts at
- * index 2 and ends at 5:
- *
- *   [[], null, ['a', 'b'], null, null, ['a'], null, []]
- *
- * @param {any} sliceClip
- * @param {"insertion"|"deletion"} type
- * @returns {GDocsRange[]}
- */
-function rangesForStyleSliceSuggestions(sliceClip, type) {
-  const text = sliceClip.resolved.dsl_spacers;
-  const locations = sliceClip.resolved[`dsl_suggested${type}s`].sgsl_sugg;
-
-  const ranges = [];
-  let openRanges = [];
-  for (let i = 0; i < locations.length; i++) {
-    let value = locations[i];
-    if (value != null) {
-      for (const range of openRanges) {
-        range.length = i - range.start;
-      }
-      openRanges = [];
-
-      for (const suggestionId of value) {
-        const range = { suggestionId, type, start: i, length: 0 };
-        ranges.push(range);
-        openRanges.push(range);
-      }
-    }
-  }
-
-  for (const range of openRanges) {
-    range.length = text.length - range.start;
-  }
-
-  return ranges;
-}
-
-/**
  * Find suggested changes in the text and mark them with appropriate DOM nodes.
  * Copying from Google Docs includes the suggestions in the text with no markup,
  * so you can't tell what's been suggested for removal or insertion. However,
@@ -795,39 +678,19 @@ function rangesForStyleSliceSuggestions(sliceClip, type) {
 function markSuggestions(node, sliceClip) {
   if (!sliceClip) return;
 
-  const rawText = sliceClip.resolved.dsl_spacers;
   const ranges = [
-    ...rangesForStyleSliceSuggestions(sliceClip, 'insertion'),
-    ...rangesForStyleSliceSuggestions(sliceClip, 'deletion'),
+    ...rangesForSuggestions(sliceClip, 'insertion'),
+    ...rangesForSuggestions(sliceClip, 'deletion'),
   ];
 
-  forEachDocsRangeInTree(
-    rawText,
-    ranges,
-    node,
-    ({ node, nodeIndex, parent, range, rangeStart }) => {
-      const rangeEnd = rangeStart + range.length;
-      const textBefore = node.value.slice(0, rangeStart);
-      const textInside = node.value.slice(rangeStart, rangeEnd);
-      const textAfter = node.value.slice(rangeEnd);
-
-      const newNodes = [];
-      if (textBefore.length) {
-        newNodes.push({ type: 'text', value: textBefore });
-        // continueIndex += 1;
-      }
-      newNodes.push({
-        type: 'element',
-        tagName: range.type === 'insertion' ? 'ins' : 'del',
-        properties: { 'data-suggestion-id': range.suggestionId },
-        children: [{ type: 'text', value: textInside }],
-      });
-      if (textAfter.length) {
-        newNodes.push({ type: 'text', value: textAfter });
-      }
-      parent.children.splice(nodeIndex, 1, ...newNodes);
-    }
-  );
+  replaceRangesInTree(sliceClipText(sliceClip), ranges, node, (range, text) => {
+    return {
+      type: 'element',
+      tagName: range.type === 'insertion' ? 'ins' : 'del',
+      properties: { 'data-suggestion-id': range.suggestionId },
+      children: [{ type: 'text', value: text }],
+    };
+  });
 }
 
 function formatSuggestions(tree, options) {
